arm64: setup: Switch over to GENERIC_CPU_DEVICES using arch_register_cpu()
To allow ACPI's _STA value to hide CPUs that are present, but not
available to online right now due to VMM of firmware policy, the
register_cpu() call needs to be made by the ACPI machinery when ACPI
is in use.

Switching to GENERIC_CPU_DEVICES is an intermediate step to allow all
four ACPI architectures to be modified at once.

Switch over to GENERIC_CPU_DEVICES, and provide an arch_register_cpu()
that populates the hotpluggable flag. arch_register_cpu() is also the
interface the ACPI machinery expects.

The struct cpu in struct cpuinfo_arm64 is never used directly, remove
it to use the one GENERIC_CPU_DEVICES provides.

Signed-off-by: James Morse <james.morse@arm.com>