ACPI: Add a new API to match a entry in the MADT Table
It is a common requirement to find an entry in the MADT table on the basis of
the ACPI ID/Processor ID. A successful match should return a MADT entry and
Physical ID corresponding to the ACPI ID.

Above match logic is required in the upcoming vcpu hotplug kernel patches.
Hence, it makes sense to refactor the existing code and extract the common
logic which could be used by current functions and upcoming code.

Signed-off-by: Salil Mehta <salil.mehta@huawei.com>