drivers: firmware: psci: Ratelimit errors printed to the console
The PSCI checker performs tests in a loop on all the CPUs. Any error
is likely to occur everytime. The current pr_err() results in pages
and pages of output. Switch these to the ratelimited version.

The current code also adds line breaks to the error messages that are
printed. While this keeps checkpatch happy, it makes grepping for those
error messages harder. Those modified by this patch are change to
all be on one line, for easier grepping.

Signed-off-by: James Morse <james.morse@arm.com>