HACK: Paper over a qemu bug with the first value of _STA for CPUs
The prototype for online-policy/hotplug is reporting CPUs as not-present
via _STA until the are first added. This is a bug as the CPUs were
descibed in the MADT, therefore they are present.

Paper over this to allow more testing.

Seen with https://github.com/salil-mehta/qemu.git
virt-cpuhp-armv8/rfc-v1-port29092022.psci.present

and commit f1247aba36e1970958b05983fbbf5a1929abe44b

Signed-off-by: James Morse <james.morse@arm.com>