net: hns3: making new refactored memcpy logic backward compatible
This patch brongs back the old-page-reuse logic which is local to the driver but
is by default disabled. This is to ensure we could still be able to upstream the
memcpy() fix patches and later we could top up the performance fixes with the
refactored fallback driver based page-reuse logic. This will clean up the driver
beside giving the performance gains.

Signed-off-by: Salil Mehta <salil.mehta@huawei.com>