net: hns3: Pass driver NAPI context to ensure Direct cache recycling
Beside the pt_ring queue, Page-pool maintains a cache for the fast recycling and
allocation of the buffers. This cache must be operated from the NAPI context of
the driver allocating/free'ng the page and hence does not requires spinlocks
unl;ike pt_ring queue. Later can reduce the perfromance in case there is a lock
contention. This helps in gain of minor performance of arounf 1-2 Gbps with
1-core 1-queue benchmaking test done.

Signed-off-by: Salil Mehta <salil.mehta@huawei.com>
Signed-off-by: Jian Shen <shenjian15@huawei.com>
Signed-off-by: Hao Chen <chenhao418@huawei.com>